packet: fix odp_packet_seg_len() casting
data_len is uint16_t, so casting the pointer to (uint32_t *) then dereference
it brings in the first 2 bytes from pkt_len.

Reviewed-and-tested-by: Maxim Uvarov <maxim.uvarov@linaro.org>

+++ platform/linux-dpdk/include/odp/packet.h
@@ -49,7 +49,7 @@ static inline uint32_t odp_packet_len(odp_packet_t pkt)
 
 static inline uint32_t odp_packet_seg_len(odp_packet_t pkt)
 {
-	return *(uint32_t *)((char *)pkt + seg_len_offset);
+	return *(uint16_t *)((char *)pkt + seg_len_offset);
 }
 
 static inline void *odp_packet_user_area(odp_packet_t pkt)
